+++ src/main/java/model/structure/Count.java
@@ -1,4 +1,4 @@
-package model.structural;
+package model.structure;
 
 import java.util.Comparator;
