+++ src/main/java/model/structure/Passage.java
@@ -1,4 +1,4 @@
-package model.structural;
+package model.structure;
 
 import java.io.Serializable;
 import java.util.*;
