+++ src/main/java/model/structure/Pitch.java
@@ -1,4 +1,4 @@
-package model.structural;
+package model.structure;
 
 import java.util.*;
